Increase Speed by Processing Image Once
Made the program process the animation image once at the beginning. I took the image and processed it into an array of textures so that draw method could just take a texture from the array instead of creating it during the game run. I increased the delay time between each cycle to keep the game from running to fast.